refactor: move disk sync logic to a composable
