+++ frontend/src/App.vue
@@ -11,14 +11,8 @@
 </template>
 
 <script setup lang="ts">
-import { inject, onMounted, onBeforeUnmount } from "vue"
 import { FrappeUIProvider } from "frappe-ui"
-import useStudioStore from "@/stores/studioStore"
+import { useDiskSync } from "@/utils/useDiskSync"
 
-const store = useStudioStore()
-const socket = inject<any>("socket")
-const onDocUpdate = (info: any) => store.applyDiskSync(info)
-
-onMounted(() => socket?.on("studio_doc_update", onDocUpdate))
-onBeforeUnmount(() => socket?.off("studio_doc_update", onDocUpdate))
+useDiskSync()
 </script>
